Added two methods: without_params and without_action.
The first one returns the url without the parameters. This come handy when implementing a function across different modules or controllers which have a differente uri segment lenght. Also may be useful on a later modification to the pagination class (for a more flexible uri_segment).

The second does the same but also removing the action. This is useful when trying to know if different functions belongs to the same controller, eg: you have a menu item 'blog' which has to be 'active' on blog/item, blog/search, etc.