Add the controller_name property to the active Request object.
When the controller class name is resolved by the router and saved to the Request
object, this will also save the controller name along with it.

For example, if the loaded controller class for the request is 'Controller_Welcome',
then the controller_name would be 'welcome'.

It can then be retrieved with the following command :
  Request::active()->controller_name