allow for created with all driver types
A few minor changes.
Completed the opening <?php tag.

Allowed the user to continue to add the database table regardless of the driver type. This will display a notification to the user, telling them to use the table they will need to set the driver type to "db".

Fixed the "migrate" tags in the help method, changed to "session".


Signed-off-by: Daniel Berry <daniel@berrymediagroup.com>