- Added ability to overwrite the options in the field completely
When using the ORM and the fieldset generation based on the Model, I
have found sometimes it is necessary to overwrite a select field's
options completely (different load conditions for the select perhaps
based on a user action).

This small tweak will allow that, but the default behaviour stays the
same.