Fix to automatic conversion of false to array in prep_vars
prep_vars is used in a number of classes to process path constants.

In php 8.1.x this automatic conversion of false to array is no
longer supported.

The false to array conversion appears to be used here to run this
processing once for the class and cache the results - using the
boolean false value of the array to indicate that the vars hadn't
been processed yet in these classes.

Fixes #2182